# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.BusinessCoinProgramStatsResponse:
    properties:
      coupons_purchased_in_period:
        type: integer
      points_received_in_period:
        type: integer
      total_coupons_purchased:
        type: integer
      total_points_received:
        type: integer
    type: object
...
  dto.BusinessRegisterRequest:
    properties:
      description:
        maxLength: 150
        type: string
      email:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      password:
        type: string
    required:
      - email
      - name
      - password
    type: object
  dto.BusinessRegisterResponse:
    properties:
      description:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  dto.BusinessReturn:
    properties:
      description:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.BusinessStatsResponse:
    properties:
      active_users:
        type: integer
      new_users:
        type: integer
      total_users:
        type: integer
    type: object
...
  dto.HTTPStatus:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  dto.RewardActivateRequest:
    properties:
      code:
        type: string
    required:
      - code
    type: object
...
  dto.RewardCreateRequest:
    properties:
      cost:
        minimum: 1
        type: integer
      description:
        maxLength: 150
        type: string
      image_url:
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
      - cost
      - image_url
      - name
    type: object
  dto.RewardCreateResponse:
    properties:
      cost:
        type: integer
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
    type: object
  dto.RewardReturn:
    properties:
      cost:
        type: integer
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
    type: object
  dto.UserEnrollCoinRequest:
    properties:
      code:
        type: string
    required:
      - code
    type: object
...
  dto.UserQRScanResponse:
    properties:
      balance:
        type: integer
      username:
        type: string
    type: object
...
host: prod-team-22-t62v97db.final.prodcontest.ru
info:
  contact: { }
  description: API of LoyalT - platform for creating and managing loyalty programs
  termsOfService: http://swagger.io/terms/
  title: LoyalT API
  version: "1.0"
paths:
...
  /business/auth/register:
    post:
      consumes:
        - application/json
      description: Register a new business in the system
      parameters:
        - description: Business registration data
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/dto.BusinessRegisterRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BusinessRegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      summary: Register new business
      tags:
        - Auth
...
  /business/coin_program/rewards:
    get:
      consumes:
        - application/json
      description: List rewards for coin program. Requires business_auth_token cookie
        for authentication.
      parameters:
        - default: 10
          description: Number of items to return
          in: query
          minimum: 1
          name: limit
          type: integer
        - default: 0
          description: Number of items to skip
          in: query
          minimum: 0
          name: offset
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: List of rewards
          schema:
            items:
              $ref: '#/definitions/dto.RewardReturn'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - BusinessAuthCookie: [ ]
      summary: List rewards
      tags:
        - Business
    post:
      consumes:
        - application/json
      description: Create new reward for coin program. Requires business_auth_token
        cookie for authentication.
      parameters:
        - description: Reward creation data
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/dto.RewardCreateRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RewardCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - BusinessAuthCookie: [ ]
      summary: Create new reward
      tags:
        - Business
  /business/coin_program/rewards/activate:
    post:
      consumes:
        - application/json
      description: Activate scanned coupon code. Requires business_auth_token cookie
        for authentication.
      parameters:
        - description: Reward activation data
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/dto.RewardActivateRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RewardReturn'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - BusinessAuthCookie: [ ]
      summary: Activate scanned coupon code
      tags:
        - Business
  /business/coin_program/scan/{code}:
    get:
      consumes:
        - application/json
      description: Scan user QR to get user data and validate QR
      parameters:
        - description: QR code data
          in: query
          name: code
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserQRScanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - BusinessAuthCookie: [ ]
      summary: Scan user QR
      tags:
        - Business
  /business/coin_program/scan/enroll:
    post:
      consumes:
        - application/json
      description: Add 1 coin to user coin program by qr code
      parameters:
        - description: User QR scan data
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/dto.UserEnrollCoinRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - BusinessAuthCookie: [ ]
      summary: Enroll coin to user by qr code
      tags:
        - Business
  /business/me:
    get:
      consumes:
        - application/json
      description: Get business profile
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BusinessReturn'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - BusinessAuthCookie: [ ]
      summary: Get business profile
      tags:
        - Auth
  /business/stats/coin_program:
    get:
      consumes:
        - application/json
      description: Get business coin program stats. Requires business_auth_token cookie
        for authentication.
      parameters:
        - description: period start date
          in: query
          name: start_date
          type: string
        - description: period end date
          in: query
          name: end_date
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BusinessCoinProgramStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - BusinessAuthCookie: [ ]
      summary: Get business coin program stats
      tags:
        - Business
...
  /business/stats/general:
    get:
      consumes:
        - application/json
      description: Get business general stats. Requires business_auth_token cookie
        for authentication.
      parameters:
        - description: period start date
          in: query
          name: start_date
          type: string
        - description: period end date
          in: query
          name: end_date
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BusinessStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HTTPStatus'
      security:
        - BusinessAuthCookie: [ ]
      summary: Get business stats
      tags:
        - Business
...
schemes:
  - https
swagger: "2.0"